fix(lens): extract the in-memory fakeStorage into one shared testing home
The synchronous Storage-over-Map fake had triplicated across activeHostStorage.test.ts, persistedPref.test.ts, and kaval/reattachAnnounce.test.ts (already diverging on a seed param). Move the seeded superset to packages/client/src/testing/fakeStorage.ts and import it in all three, deleting the local copies.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-3, raised by hickey). Not pushed or merged.
